Trying to setup a breadboard test for Collector Plate
