+++ net_maps/mil_col_net_map.csv
@@ -1,4 +1,34 @@
 Board,Net,Component,Designator,Connector Name,,
-Millan,HIL_OUT,breadboard,1,FRED,,
-Millan,HIL_AIN,breadboard,2,FRED,,
-Millan,HIL_DIN,breadboard,3,FRED,,
+Collector,MUX_A,J4,2,ABOX,,
+Collector,MUX_B,J4,1,ABOX,,
+Collector,MUX_C,J4,4,ABOX,,
+Collector,MUX_D,J4,7,ABOX,,
+Collector,TEMP_OUT,J4,5,ABOX,,
+
+Collector,GND_I,J4,3,ABOX,,
+
+Collector,3V3,J4,8,ABOX,,
+Collector,CELL_0,J3,1,ORION,,
+Collector,CELL_2,J3,2,ORION,,
+Collector,CELL_4,J3,3,ORION,,
+Collector,CELL_6,J3,4,ORION,,
+Collector,CELL_8,J3,5,ORION,,
+Collector,CELL_10,J3,6,ORION,,
+Collector,CELL_12,J3,7,ORION,,
+Collector,CELL_13,J3,8,ORION,,
+Collector,CELL_15,J3,9,ORION,,
+Collector,CELL_17,J3,10,ORION,,
+Collector,CELL_19,J3,11,ORION,,
+Collector,CELL_19,J3,12,ORION,,
+Collector,CELL_1,J3,13,ORION,,
+Collector,CELL_3,J3,14,ORION,,
+Collector,CELL_5,J3,15,ORION,,
+Collector,CELL_7,J3,16,ORION,,
+Collector,CELL_9,J3,17,ORION,,
+Collector,CELL_11,J3,18,ORION,,
+Collector,CELL_12,J3,19,ORION,,
+Collector,CELL_14,J3,20,ORION,,
+Collector,CELL_16,J3,21,ORION,,
+Collector,CELL_18,J3,22,ORION,,
+Collector,CELL_19,J3,23,ORION,,
+Collector,CELL_19,J3,24,ORION,,
