+++ scripts/test_mil_col.py
@@ -8,36 +8,50 @@ import time
 
 
 # ---------------------------------------------------------------------------- #
-def same_to_color_str(same: bool) -> str:
-	if same:
-		return utils.bcolors.OKGREEN + "SUCCESS" + utils.bcolors.ENDC
-	else:
-		return utils.bcolors.FAIL + "FAILURE" + utils.bcolors.ENDC
+def test_collector(hil: HIL):
+	# HIL ouputs -> Collector inputs
+	mux_a = hil.dout("Collector", "MUX_A")
+	mux_b = hil.dout("Collector", "MUX_B")
+	mux_c = hil.dout("Collector", "MUX_C")
+	mux_d = hil.dout("Collector", "MUX_D")
 
+	# HIL inputs -> Collector outputs
+	temp_out = hil.ain("Collector", "TEMP_OUT")
 
-def test(hil: HIL):
-	hil_out = hil.dout("Millan", "HIL_OUT")
-	hil_ain = hil.ain("Millan", "HIL_AIN")
-	hil_din = hil.din("Millan", "HIL_DIN")
+	tolerance_v    = 0.1 # volts
+	current_res    = 9100.0 # ohms
+	pullup_res     = 4700.0 # ohms
+	test_voltage   = 3.3 # volts
+	pullup_voltage = 5 # volts
+	num_therm      = 10
 
-	for _i in range(3):
-		for state in [0, 1]:
-			print(f"\nHIL_OUT: {state}")
-			hil_out.state = state
-			time.sleep(0.5)
+	test_voltage = (pullup_voltage / (current_res + pullup_res)) * current_res
+	utils.log_warning(f"Test voltage: {test_voltage}")
+	
 
-			hil_din_state = hil_din.state
-			din_same = hil_din_state == state
-			print(f"HIL_DIN: {hil_din_state} == {state} -> {same_to_color_str(din_same)}")
+	for thermistor in range(num_therm):
+		print(f"\nPlace test input on thermistor {thermistor}.")
+		input("Press Enter when ready...")
 
-			hil_ain_state = hil_ain.state
-			ain_target = state * 5.0
-			ain_same = abs(hil_ain_state - ain_target) < 0.1
-			print(f"HIL_AIN: {hil_ain_state} == {ain_target} -> {same_to_color_str(ain_same)}")
+		for i in range(num_therm):
+			# MUX (multiplexer) = choose which output to return from the thermistor based on the input
+			# Like a giant switch statement (0 -> return thermistor 0, 1 -> return thermistor 1, etc.)
+			# Encode the current thermistor into binary where each bit corresponds to each pin being high or low
+			mux_a.state = i & 0x1
+			mux_b.state = i & 0x2
+			mux_c.state = i & 0x4
+			mux_d.state = i & 0x8
+			time.sleep(0.01)
 
-			time.sleep(1)
+			temp_out_state = temp_out.state
+			if i == thermistor: expected_voltage = test_voltage
+			else:               expected_voltage = pullup_voltage
+			within = abs(temp_out_state - expected_voltage) < tolerance_v
+			
+			if within: within_text = utils.bcolors.OKGREEN + "PASS" + utils.bcolors.ENDC
+			else:      within_text = utils.bcolors.FAIL + "FAIL" + utils.bcolors.ENDC
 
-		print()
+			print(f"({thermistor=}, {i=})  temp_out_state={temp_out_state:.1f} ?= expected_voltage={expected_voltage:.1f}  ->  {within_text}")
 # ---------------------------------------------------------------------------- #
 
 
@@ -45,9 +59,9 @@ def test(hil: HIL):
 if __name__ == "__main__":
 	hil = HIL()
 
-	hil.load_config("config_millan.json")
-	hil.load_pin_map("millan_net_map.csv", "stm32f407_pin_map.csv")
+	hil.load_config("config_mil_col.json")
+	hil.load_pin_map("mil_col_net_map.csv", "stm32f407_pin_map.csv")
 	
-	test(hil)
+	test_collector(hil)
 
 	hil.shutdown()
